Move multi dimensional sets to samples
