+++ smolyay/__init__.py
@@ -1,3 +1,2 @@
 from . import basis
 from . import samples
-from . import grid
